Rename home page component
